# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Tabdeel - Barter Exchange Platform API for Jordan
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /swap-requests:
    get:
      operationId: getSwapRequests
      tags: [swapRequests]
      summary: Get my swap requests (sent and received)
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [sent, received]
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, accepted, rejected, completed]
      responses:
        "200":
          description: Swap requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SwapRequest"
    post:
      operationId: createSwapRequest
      tags: [swapRequests]
      summary: Request a swap on a listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SwapRequestInput"
      responses:
        "201":
          description: Swap request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SwapRequest"
  /swap-requests/{id}/accept:
    patch:
      operationId: acceptSwapRequest
      tags: [swapRequests]
      summary: Accept a swap request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SwapRequest"
  /swap-requests/{id}/reject:
    patch:
      operationId: rejectSwapRequest
      tags: [swapRequests]
      summary: Reject a swap request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SwapRequest"
  /swap-requests/{id}/complete:
    patch:
      operationId: completeSwapRequest
      tags: [swapRequests]
      summary: Mark swap as completed
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SwapRequest"

  # ─── MESSAGES ───────────────────────────────────────────────────────────────
...
  /conversations/{swapRequestId}/messages:
    get:
      operationId: getMessages
      tags: [messages]
      summary: Get messages in a conversation
      parameters:
        - name: swapRequestId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message
      parameters:
        - name: swapRequestId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /conversations/{swapRequestId}/read:
    patch:
      operationId: markConversationRead
      tags: [messages]
      summary: Mark all messages in conversation as read
      parameters:
        - name: swapRequestId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked as read

  # ─── RATINGS ────────────────────────────────────────────────────────────────
...
  /reports:
    post:
      operationId: createReport
      tags: [reports]
      summary: Report a listing or user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "201":
          description: Report submitted

  # ─── NOTIFICATIONS ──────────────────────────────────────────────────────────
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked as read

  # ─── IMAGE UPLOAD ────────────────────────────────────────────────────────────
...
  /admin/reports:
    get:
      operationId: getAdminReports
      tags: [admin]
      summary: Get all reports
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, resolved, dismissed]
      responses:
        "200":
          description: Reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
  /admin/reports/{id}:
    patch:
      operationId: resolveReport
      tags: [admin]
      summary: Resolve or dismiss a report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolveReportInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"

# ─── SCHEMAS ────────────────────────────────────────────────────────────────
components:
  schemas:
...
    UserProfile:
      type: object
      required: [id, name, city, avatarUrl, rating, ratingCount, listingsCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        city:
          type: string
        avatarUrl:
          type: ["string", "null"]
        rating:
          type: number
        ratingCount:
          type: integer
        listingsCount:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    Category:
      type: object
      required: [id, name, slug, icon]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        icon:
          type: string
        listingsCount:
          type: integer

    # LISTINGS
    Listing:
      type: object
      required: [id, title, description, condition, wantsInExchange, city, images, category, owner, isFeatured, isActive, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        condition:
          type: string
          enum: [new, used, good]
        wantsInExchange:
          type: string
        city:
          type: string
        images:
          type: array
          items:
            type: string
        category:
          $ref: "#/components/schemas/Category"
        owner:
          $ref: "#/components/schemas/UserProfile"
        isFeatured:
          type: boolean
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    SwapRequest:
      type: object
      required: [id, listing, requester, status, createdAt]
      properties:
        id:
          type: integer
        listing:
          $ref: "#/components/schemas/Listing"
        requester:
          $ref: "#/components/schemas/UserProfile"
        message:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, accepted, rejected, completed]
        createdAt:
          type: string
          format: date-time
    SwapRequestInput:
      type: object
      required: [listingId]
      properties:
        listingId:
          type: integer
        message:
          type: string

    # MESSAGES
    Message:
      type: object
      required: [id, swapRequestId, sender, content, isRead, createdAt]
      properties:
        id:
          type: integer
        swapRequestId:
          type: integer
        sender:
          $ref: "#/components/schemas/UserProfile"
        content:
          type: string
        imageUrl:
          type: ["string", "null"]
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
        imageUrl:
          type: string
...
    Report:
      type: object
      required: [id, reporter, targetType, targetId, reason, status, createdAt]
      properties:
        id:
          type: integer
        reporter:
          $ref: "#/components/schemas/UserProfile"
        targetType:
          type: string
          enum: [listing, user]
        targetId:
          type: integer
        reason:
          type: string
        status:
          type: string
          enum: [pending, resolved, dismissed]
        createdAt:
          type: string
          format: date-time
    ReportInput:
      type: object
      required: [targetType, targetId, reason]
      properties:
        targetType:
          type: string
          enum: [listing, user]
        targetId:
          type: integer
        reason:
          type: string
    ResolveReportInput:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [resolved, dismissed]

    # NOTIFICATIONS
...