# lib/api-zod/src/generated/types/resolveReportInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Tabdeel - Barter Exchange Platform API for Jordan
 * OpenAPI spec version: 0.1.0
 */
import type { ResolveReportInputStatus } from './resolveReportInputStatus';

export interface ResolveReportInput {
  status: ResolveReportInputStatus;
}